Palette "Hotfix"
Huzzah! The world shall be blessed, and will now not be contained to only black and white. It now possesses a retro but crisp Pico8 palette.
After a not-so-quick revision of the whole ColorPicker system, I found out a few errors and fixed them. Namely, the DebugText object wasn't dragged into the "Debug Text" field of the Color Picker object.
Another error was with the delta that checks if the position of the mouse is inside of the palette texture. To be more specific, a line of code needed to be changed from "float y = Mathf.Clamp(delta.y / height, 1f, 0f);" to "float y = Mathf.Clamp(delta.y / height, 0f, 1f);".
Background is also fixed. I just made a new material for it, changed the shading to "Unlit", so now it's a blinding white :) .
Added a new file named "todo.txt", just tracking what comes up next.