Added a button to make a new frame
I guess this is the beginning of the code organization process. There are a lot of functions in the LineGenerator script that could be separated for clarity.
The new frame button was a bit tricky to add, but nothing too bad. I also spiced up the font, making it a tad wider and funkier.

mini rant
I presented the app to some people and the reactions were... unsurprising. They were a bit impressed by the state that the app was in, but they also requested multiple features, and I think that a lot of people are missing the whole point of these types of projects. It's not about making a finished product and releasing it as a competitor to the big names. It's not about the end point, it's about the journey. It's about how far you've managed to get. It's about reverse engineering and figuring out why things are the way they are. It's about saying "I made this. I. I managed to make this, and I understand it". Maybe I'm rambling. Maybe.
rant over

ok uhhhhhhhhh enjoy this commit and see you next time when we will hopefully have gifs or better UI or other things.