Check-up
Just checking if everything works accordingly. I realized a little problem with the gif compiler and how files are sorted (for a computer, the sequence of strings "1" "2" "3" "4" "5" "6" "7" "8" "9" "10" "11" "12" is not sorted, and it takes them like "1" "10" "11" "12" "2" "3" "4" and so on. I fixed it by adding two zeroes in front of the number if the number is lower than 10, one zero if the number is higher than 10 but lower than 100. This way it'll respect the order until the 1000th frame, and I doubt any animation on this app will feature that many frames